Create MacOS/Linux compatibility [NOT TESTED]
created install.sh and run.sh so that the app can be run with a Mac or Linux computer.

Please note; that the implementation is not yet tested (don't have a mac) and is not guaranteed to work.
Will hopefully have it tested before the tomorrow's meeting.
"test-status"  in install.sh and run.sh will be changed once the file is tested